software : get rid of float variables - #109
it has been recommended to use only integer values.

+++ src/software/firmware/includes/pression.h
@@ -18,4 +18,4 @@
  * used when in simulation mode)
  * @return         The current pressure in mmH20
  */
-int readPressureSensor(uint16_t centiSec);
+int16_t readPressureSensor(uint16_t centiSec);
